Rfc/issue 355 organize serve lifecycle (#419)
* task: organize serve

* fix: remove ctx from resolve

* fix: refactor further

* task: scope filters by file
